Call Session#onInventoryClose when the plugin is disabled.
Not all changes in this are required but were, along with the previous commit, intended to address the issue where players lose the contents of an open MassStorage chest when the plugin is disabled, for example during a server restart. The first scapegoat was async saving, the second was failing to actually close the inventory.
As it turns out, the real cause of the trouble was that the EventHandler for InventoryCloseEvent was no longer notified if the closing was triggered in the JavaPlugin#onDisable method. I'm guessing that all Listeners get unregistered before this is called. Therefore, a manual call to the handler is necessary, and the problem was fixed.