Require the entire and use pattern matching as expression
When parts of the searched categories are accepted, we get noisy results.  For example, searching for "food" also yields several flower types, for unknown reasons.  It is reasonable to assume that players want to enter the entire cateegory they are looking for, as opposed to matching categories.

The switch case was used without a return value but with side effects.  This commit uses it as an expression.